+++ react/components/WordpressRelatedPostsBlock.tsx
@@ -22,6 +22,7 @@ const WordpressRelatedPostsBlock: StorefrontFunctionComponent<WPRelatedPostsBloc
   showDates,
   showAuthors,
   showExcerpts,
+  absoluteLinks,
   numberOfPosts,
   customDomain,
   customDomainSlug,
@@ -61,6 +62,7 @@ const WordpressRelatedPostsBlock: StorefrontFunctionComponent<WPRelatedPostsBloc
                     date={post.date}
                     id={post.id}
                     slug={post.slug}
+                    link={post.link}
                     author={post.author ? post.author.name : ''}
                     excerpt={post.excerpt.rendered}
                     category={post.categories[0]?.name ?? ''}
@@ -74,6 +76,7 @@ const WordpressRelatedPostsBlock: StorefrontFunctionComponent<WPRelatedPostsBloc
                     showDate={showDates}
                     showAuthor={showAuthors}
                     showExcerpt={showExcerpts}
+                    absoluteLinks={absoluteLinks}
                     useTextOverlay={useTextOverlays}
                   />
                 </div>
@@ -96,6 +99,7 @@ interface WPRelatedPostsBlockProps {
   showDates: boolean
   showAuthors: boolean
   showExcerpts: boolean
+  absoluteLinks: boolean
   productQuery: ProductQuery
   customDomain: string
   customDomainSlug: string
@@ -165,6 +169,7 @@ WordpressRelatedPostsBlock.defaultProps = {
   showDates: true,
   showAuthors: false,
   showExcerpts: false,
+  absoluteLinks: false,
   customDomain: undefined,
   customDomainSlug: undefined,
 }
@@ -234,6 +239,14 @@ const messages = defineMessages({
     defaultMessage: '',
     id: 'admin/editor.wordpressExcerpts.description',
   },
+  absoluteLinksTitle: {
+    defaultMessage: '',
+    id: 'admin/editor.absoluteLinks.title',
+  },
+  absoluteLinksDescription: {
+    defaultMessage: '',
+    id: 'admin/editor.absoluteLinks.description',
+  },
   customDomainTitle: {
     defaultMessage: '',
     id: 'admin/editor.wordpressCustomDomain.title',
@@ -302,6 +315,13 @@ WordpressRelatedPostsBlock.schema = {
     showExcerpts: {
       title: messages.showExcerptsTitle.id,
       description: messages.showExcerptsDescription.id,
+      type: 'boolean',
+      isLayout: false,
+      default: false,
+    },
+    absoluteLinks: {
+      title: messages.absoluteLinksTitle.id,
+      description: messages.absoluteLinksDescription.id,
       type: 'boolean',
       isLayout: false,
       default: false,