+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `absoluteLinks` prop to some components, to allow linking to blog posts on external WP site
+
 ## [2.0.1] - 2020-07-24
 
 ### Added